added Cart and CartLineItem models

+++ server/db/models/OrderLineItem.js
@@ -3,7 +3,7 @@ const { db } = require('../index');
 
 const OrderLineItem = db.define('order_line_item', {
   unitPrice: {
-    type: Sequelize.FLOAT(99999, 2),
+    type: Sequelize.DECIMAL(10, 2),
     allowNull: false,
   },
   quantity: {
@@ -11,7 +11,7 @@ const OrderLineItem = db.define('order_line_item', {
     allowNull: false,
   },
   totalPrice: {
-    type: Sequelize.FLOAT(99999, 2),
+    type: Sequelize.DECIMAL(10, 2),
   },
 });
